refactor: restructured the project setup
