--- devs-who-run-api/Modules/Auth/Services/Interfaces/IAuthService.cs
@@ -1,11 +1,0 @@
-using devs_who_run_api.Modules.Auth.DTOs;
-
-namespace devs_who_run_api.Modules.Auth.Services.Interfaces;
-
-public interface IAuthService
-{
-    Task<AuthResponse> RegisterAsync(RegisterRequest request);
-    Task<AuthResponse> LoginAsync(LoginRequest request);
-    Task<AuthResponse> RefreshTokenAsync(string refreshToken);
-    Task<AuthResponse> GitHubLoginAsync(string code);
-}
