+++ src/DevsWhoRun.Api/DevsWhoRunDBContext.cs
@@ -1,0 +1,8 @@
+using Microsoft.EntityFrameworkCore;
+
+namespace DevsWhoRun.Api;
+
+public class DevsWhoRunDbContext(DbContextOptions<DevsWhoRunDbContext> options) : DbContext(options)
+{
+    public DbSet<Member> Members { get; set; }
+}
